Move @Generated annotation from gRPC to core
