+++ extensions/grpc/codegen/src/main/java/io/quarkus/grpc/deployment/GrpcPostProcessing.java
@@ -25,7 +25,7 @@ public class GrpcPostProcessing {
     private static final String POST_PROCESS_NO_FINAL = "quarkus.generate-code.grpc-post-processing.no-final";
     // this is intentionally split so that it doesn't get replaced by the Jakarta transformer
     public static final String JAVAX_GENERATED = "javax" + ".annotation.Generated";
-    public static final String QUARKUS_GENERATED = "io.quarkus.grpc.common.Generated";
+    public static final String QUARKUS_GENERATED = "io.quarkus.Generated";
     public static final String STUB = "Stub";
     public static final String BIND_METHOD = "bindService";
